+++ src/OpenTelemetry/Logs/Pool/LogRecordPoolHelper.cs
@@ -10,6 +10,8 @@ internal static class LogRecordPoolHelper
 
     public static void Clear(LogRecord logRecord)
     {
+        logRecord.Dropped = false;
+
         var attributeStorage = logRecord.AttributeStorage;
         if (attributeStorage != null)
         {